Relationship setup| clean scopes from Model to Trait
